+++ src/oms_sensemaking/object_minimums/sensemaker.py
@@ -4,51 +4,16 @@ import logging
 
 from oms_sdk.generated.generated_graphql_client import (
     AttributeAttribute,
-    AttributeQuery,
-    NodeNode,
-    RelationshipNodeQuery,
-    RelationshipQuery,
+    NodeQuery,
+    RelationshipRelationship,
 )
 
 from oms_sensemaking.core.oms_crud import OmsCrudTool
 from oms_sensemaking.core.sensemakers import Sensemaker
+from oms_sensemaking.object_minimums.object_minimum_data_retriever import ObjectMinimumDataRetriever
 from oms_sensemaking.object_minimums.object_minimum_models import ObjectMinimumRubric
 
 LOGGER = logging.getLogger(__name__)
-
-
-class ObjectMinimumDataRetriever:
-    def retrieve_data_for_grading(
-        self,
-        oms_crud_tool: OmsCrudTool,
-        node: NodeNode,
-        required_attributes: list[str],
-        required_relationships: list[str],
-    ):
-        iris_to_grade = {"attributes": None, "relationships": None}
-        if required_attributes:
-            try:
-                attributes_to_grade = oms_crud_tool.get_attributes(
-                    AttributeQuery(nodeIds=[node.id], attributeIris=required_attributes)
-                )
-                iris_to_grade["attributes"] = attributes_to_grade
-            except Exception as e:
-                LOGGER.error("Error retrieving attributes for node ID %s: %s", node.id, str(e))
-                raise
-
-        if required_relationships:
-            try:
-                relationships_to_grade = oms_crud_tool.get_relationships(
-                    RelationshipQuery(
-                        nodes=RelationshipNodeQuery(nodeIds=[node.id]), objectPropertyIris=required_relationships
-                    )
-                )
-                iris_to_grade["relationships"] = relationships_to_grade
-            except Exception as e:
-                LOGGER.error("Error retrieving relationships for node ID %s: %s", node.id, str(e))
-                raise
-
-        return iris_to_grade
 
 
 class ObjectMinimums(Sensemaker):
@@ -80,7 +45,11 @@ class ObjectMinimums(Sensemaker):
         self.obj_min_rubric = obj_min_rubric
         self.obj_min_retriever = obj_min_retriever
 
-    def process_data(self, attribute_of_node: AttributeAttribute, config: dict | None = None):
+    def process_data(
+        self,
+        node_characteristic: AttributeAttribute | RelationshipRelationship,
+        config: dict | None = None,
+    ):
         """
         Determine the completeness of an ATOMS node based off its various related components
 
@@ -88,50 +57,55 @@ class ObjectMinimums(Sensemaker):
         other descriptive objects in the future)
         :return: Grade
         """
-        LOGGER.info("Running Object Minimums Sensemaker")
-
-        """
-        possible structure:
-
-        Probably get the node object by doing: node = crud_tool.get_node(attribute_of_node.nodeId)
-
-        Look at the node obj class iri: iri = node.iri
-
-        call some sort of helper function(s) below to call the grading class
-
-        run the grading rubric made for the specific class iri on the node
-
-        retrieve the grade
-
-        pass the grade to another helper to push the grade info to
-        the node's metadata (discussed with effects team) is this an update? (unsure)
-        """
-
         try:
-            try:
-                node_id = attribute_of_node.nodeId
-                node = self.oms_crud_tool.get_node(node_id)
-            except Exception as e:
-                LOGGER.error("Unexpected error retrieving node ID %s: %s", node_id, str(e))
-                raise
-            class_iri = node.classIri
-            required_attributes, required_relationships = self._get_required_iris(class_iri)
-            required_iris = required_attributes + required_relationships
+            # attribute has only one node it is associated with, relationships
+            # have a start node and end node. hard to tell what exactly is the node
+            # we care about, so let's just care for both
 
-            if not required_iris:
-                LOGGER.info("Object Minimum grade not calculated due to lack of required IRIs")
-                return []
+            """
+            what I want to do here in the future is to check if it is attr and rel
+            and call a custom class/function to retr data all at once with a custom query
+
+            get node(s)
+            get node(s) attrs and rels
+            send the data through helper functions
+
+            this will allow us to eliminate the .retrieve_data_for_grading()
+            or instead move it up/modify it to grab all data all at once
+            """
+            if isinstance(node_characteristic, AttributeAttribute):
+                node_ids = [node_characteristic.nodeId]
+            elif isinstance(node_characteristic, RelationshipRelationship):
+                node_ids = [node_characteristic.startNodeId, node_characteristic.endNodeId]
             else:
-                self.obj_min_rubric.required_iris = required_iris
+                LOGGER.warning(
+                    "Unexpected characteristic type %s in ObjectMinimums",
+                    type(node_characteristic),
+                )
+                return []
+            node_query = NodeQuery(ids=node_ids)
+            nodes = self.oms_crud_tool.get_nodes(node_query)
+            for node in nodes.data:
+                required_attributes, required_relationships = self._get_required_iris(node.classIri)
+                required_iris = required_attributes + required_relationships
 
-            iris_to_grade = self.obj_min_retriever.retrieve_data_for_grading(
-                self.oms_crud_tool, node, required_attributes, required_relationships
-            )
-            grade = self._calculate_grade(iris_to_grade["attributes"], iris_to_grade["relationships"])
+                if not required_iris:
+                    LOGGER.info(
+                        "Ungradeable class object. Grade was not able to be calculated for the node with id: %s",
+                        node.id,
+                    )
+                    continue
+                else:
+                    self.obj_min_rubric.required_iris = required_iris
 
-            LOGGER.info("Object Minimum grade: %s", grade.completion_score)
+                retrieved_node_data = self.obj_min_retriever.retrieve_data_for_grading(
+                    self.oms_crud_tool, node, required_attributes, required_relationships
+                )
+                grade = self._calculate_grade(retrieved_node_data["attributes"], retrieved_node_data["relationships"])
+                # TODO: Update the node metadata with grade (amongst other various fields) once schema support exists
+                LOGGER.info("Object Minimum grade: %s", grade.completion_score)
         except Exception as e:
-            LOGGER.error("Error processing object minimum data for node ID %s: %s", attribute_of_node.nodeId, str(e))
+            LOGGER.error("Error processing object minimum data for object(s) %s: %s", node_ids, str(e))
 
         return []